Tolerate toxic containers
A toxic container is one which cannot be deleted. These occur with
monotonous regularity during development.

Rather than fail to start a container when a toxic container cannot be
destroyed, diagnose the problem by logging, but then continue. This is
less strict, but more failure tolerant and therefore more robust
in production.